Add tags to endpoints. Restructure api/nodes.py so that we have one file with only endpoint routes and all helper methods go into the separate internal module
