+++ datajunction-server/datajunction_server/api/engines.py
@@ -12,7 +12,7 @@ from datajunction_server.api.helpers import get_engine
 from datajunction_server.models.engine import Engine, EngineInfo
 from datajunction_server.utils import get_session
 
-router = APIRouter()
+router = APIRouter(tags=["engines"])
 
 
 @router.get("/engines/", response_model=List[EngineInfo])
@@ -33,14 +33,19 @@ def get_an_engine(
     return get_engine(session, name, version)
 
 
-@router.post("/engines/", response_model=EngineInfo, status_code=201)
-def add_an_engine(
+@router.post(
+    "/engines/",
+    response_model=EngineInfo,
+    status_code=201,
+    name="Add An Engine",
+)
+def add_engine(
     data: EngineInfo,
     *,
     session: Session = Depends(get_session),
 ) -> EngineInfo:
     """
-    Add an Engine
+    Add a new engine
     """
     try:
         get_engine(session, data.name, data.version)
