+++ datajunction-server/datajunction_server/api/metrics.py
@@ -17,7 +17,7 @@ from datajunction_server.models.node import DimensionAttributeOutput, Node, Node
 from datajunction_server.sql.dag import get_shared_dimensions
 from datajunction_server.utils import get_session
 
-router = APIRouter()
+router = APIRouter(tags=["metrics"])
 
 
 def get_metric(session: Session, name: str) -> Node:
